+++ src/composer/compiler.rs
@@ -117,8 +117,8 @@ impl Compiler {
         let mut q_l = vec![BlsScalar::zero(); size];
         let mut q_r = vec![BlsScalar::zero(); size];
         let mut q_o = vec![BlsScalar::zero(); size];
+        let mut q_4 = vec![BlsScalar::zero(); size];
         let mut q_c = vec![BlsScalar::zero(); size];
-        let mut q_d = vec![BlsScalar::zero(); size];
         let mut q_arith = vec![BlsScalar::zero(); size];
         let mut q_range = vec![BlsScalar::zero(); size];
         let mut q_logic = vec![BlsScalar::zero(); size];
@@ -130,8 +130,8 @@ impl Compiler {
             q_l[i] = c.q_l;
             q_r[i] = c.q_r;
             q_o[i] = c.q_o;
+            q_4[i] = c.q_4;
             q_c[i] = c.q_c;
-            q_d[i] = c.q_d;
             q_arith[i] = c.q_arith;
             q_range[i] = c.q_range;
             q_logic[i] = c.q_logic;
@@ -143,8 +143,8 @@ impl Compiler {
         let q_l_poly = domain.ifft(&q_l);
         let q_r_poly = domain.ifft(&q_r);
         let q_o_poly = domain.ifft(&q_o);
+        let q_4_poly = domain.ifft(&q_4);
         let q_c_poly = domain.ifft(&q_c);
-        let q_d_poly = domain.ifft(&q_d);
         let q_arith_poly = domain.ifft(&q_arith);
         let q_range_poly = domain.ifft(&q_range);
         let q_logic_poly = domain.ifft(&q_logic);
@@ -155,8 +155,8 @@ impl Compiler {
         let q_l_poly = Polynomial::from_coefficients_vec(q_l_poly);
         let q_r_poly = Polynomial::from_coefficients_vec(q_r_poly);
         let q_o_poly = Polynomial::from_coefficients_vec(q_o_poly);
+        let q_4_poly = Polynomial::from_coefficients_vec(q_4_poly);
         let q_c_poly = Polynomial::from_coefficients_vec(q_c_poly);
-        let q_d_poly = Polynomial::from_coefficients_vec(q_d_poly);
         let q_arith_poly = Polynomial::from_coefficients_vec(q_arith_poly);
         let q_range_poly = Polynomial::from_coefficients_vec(q_range_poly);
         let q_logic_poly = Polynomial::from_coefficients_vec(q_logic_poly);
@@ -173,8 +173,8 @@ impl Compiler {
         let q_l_poly_commit = commit_key.commit(&q_l_poly).unwrap_or_default();
         let q_r_poly_commit = commit_key.commit(&q_r_poly).unwrap_or_default();
         let q_o_poly_commit = commit_key.commit(&q_o_poly).unwrap_or_default();
+        let q_4_poly_commit = commit_key.commit(&q_4_poly).unwrap_or_default();
         let q_c_poly_commit = commit_key.commit(&q_c_poly).unwrap_or_default();
-        let q_d_poly_commit = commit_key.commit(&q_d_poly).unwrap_or_default();
         let q_arith_poly_commit =
             commit_key.commit(&q_arith_poly).unwrap_or_default();
         let q_range_poly_commit =
@@ -199,8 +199,8 @@ impl Compiler {
             q_l: q_l_poly_commit,
             q_r: q_r_poly_commit,
             q_o: q_o_poly_commit,
+            q_4: q_4_poly_commit,
             q_c: q_c_poly_commit,
-            q_4: q_d_poly_commit,
             q_arith: q_arith_poly_commit,
         };
 
@@ -252,8 +252,8 @@ impl Compiler {
             q_l: q_l_poly,
             q_r: q_r_poly,
             q_o: q_o_poly,
+            q_4: q_4_poly,
             q_c: q_c_poly,
-            q_4: q_d_poly,
             q_arith: q_arith_poly,
             q_range: q_range_poly,
             q_logic: q_logic_poly,
@@ -343,8 +343,8 @@ impl Compiler {
             q_l: (selectors.q_l.clone(), q_l_eval_8n.clone()),
             q_r: (selectors.q_r.clone(), q_r_eval_8n.clone()),
             q_o: (selectors.q_o, q_o_eval_8n),
-            q_c: (selectors.q_c.clone(), q_c_eval_8n.clone()),
             q_4: (selectors.q_4, q_4_eval_8n),
+            q_c: (selectors.q_c.clone(), q_c_eval_8n.clone()),
             q_arith: (selectors.q_arith, q_arith_eval_8n),
         };