+++ src/proof_system/widget/permutation/proverkey.rs
@@ -48,7 +48,7 @@ impl ProverKey {
         index: usize,
         a_w_i: &BlsScalar,
         b_w_i: &BlsScalar,
-        c_w_i: &BlsScalar,
+        o_w_i: &BlsScalar,
         d_w_i: &BlsScalar,
         z_i: &BlsScalar,
         z_i_next: &BlsScalar,
@@ -58,22 +58,22 @@ impl ProverKey {
         gamma: &BlsScalar,
     ) -> BlsScalar {
         let a = self.compute_quotient_identity_range_check_i(
-            index, a_w_i, b_w_i, c_w_i, d_w_i, z_i, alpha, beta, gamma,
+            index, a_w_i, b_w_i, o_w_i, d_w_i, z_i, alpha, beta, gamma,
         );
         let b = self.compute_quotient_copy_range_check_i(
-            index, a_w_i, b_w_i, c_w_i, d_w_i, z_i_next, alpha, beta, gamma,
+            index, a_w_i, b_w_i, o_w_i, d_w_i, z_i_next, alpha, beta, gamma,
         );
         let c = self.compute_quotient_term_check_one_i(z_i, l1_alpha_sq);
         a + b + c
     }
-    // (a(x) + beta * X + gamma) (b(X) + beta * k1 * X + gamma) (c(X) + beta *
+    // (a(x) + beta * X + gamma) (b(X) + beta * k1 * X + gamma) (o(X) + beta *
     // k2 * X + gamma)(d(X) + beta * k3 * X + gamma)z(X) * alpha
     fn compute_quotient_identity_range_check_i(
         &self,
         index: usize,
         a_w_i: &BlsScalar,
         b_w_i: &BlsScalar,
-        c_w_i: &BlsScalar,
+        o_w_i: &BlsScalar,
         d_w_i: &BlsScalar,
         z_i: &BlsScalar,
         alpha: &BlsScalar,
@@ -84,12 +84,12 @@ impl ProverKey {
 
         (a_w_i + (beta * x) + gamma)
             * (b_w_i + (beta * K1 * x) + gamma)
-            * (c_w_i + (beta * K2 * x) + gamma)
+            * (o_w_i + (beta * K2 * x) + gamma)
             * (d_w_i + (beta * K3 * x) + gamma)
             * z_i
             * alpha
     }
-    // (a(x) + beta* Sigma1(X) + gamma) (b(X) + beta * Sigma2(X) + gamma) (c(X)
+    // (a(x) + beta* Sigma1(X) + gamma) (b(X) + beta * Sigma2(X) + gamma) (o(X)
     // + beta * Sigma3(X) + gamma)(d(X) + beta * Sigma4(X) + gamma) Z(X.omega) *
     // alpha
     fn compute_quotient_copy_range_check_i(
@@ -97,7 +97,7 @@ impl ProverKey {
         index: usize,
         a_w_i: &BlsScalar,
         b_w_i: &BlsScalar,
-        c_w_i: &BlsScalar,
+        o_w_i: &BlsScalar,
         d_w_i: &BlsScalar,
         z_i_next: &BlsScalar,
         alpha: &BlsScalar,
@@ -111,7 +111,7 @@ impl ProverKey {
 
         let product = (a_w_i + (beta * s_sigma_1_eval) + gamma)
             * (b_w_i + (beta * s_sigma_2_eval) + gamma)
-            * (c_w_i + (beta * s_sigma_3_eval) + gamma)
+            * (o_w_i + (beta * s_sigma_3_eval) + gamma)
             * (d_w_i + (beta * s_sigma_4_eval) + gamma)
             * z_i_next
             * alpha;
@@ -131,7 +131,7 @@ impl ProverKey {
         &self,
         z_challenge: &BlsScalar,
         (alpha, beta, gamma): (&BlsScalar, &BlsScalar, &BlsScalar),
-        (a_eval, b_eval, c_eval, d_eval): (
+        (a_eval, b_eval, o_eval, d_eval): (
             &BlsScalar,
             &BlsScalar,
             &BlsScalar,
@@ -146,13 +146,13 @@ impl ProverKey {
         z_poly: &Polynomial,
     ) -> Polynomial {
         let a = self.compute_linearizer_identity_range_check(
-            (a_eval, b_eval, c_eval, d_eval),
+            (a_eval, b_eval, o_eval, d_eval),
             z_challenge,
             (alpha, beta, gamma),
             z_poly,
         );
         let b = self.compute_linearizer_copy_range_check(
-            (a_eval, b_eval, c_eval),
+            (a_eval, b_eval, o_eval),
             z_eval,
             sigma_1_eval,
             sigma_2_eval,
@@ -172,10 +172,10 @@ impl ProverKey {
         &(&a + &b) + &c
     }
     // (a_eval + beta * z_challenge + gamma)(b_eval + beta * K1 * z_challenge +
-    // gamma)(c_eval + beta * K2 * z_challenge + gamma) * alpha z(X)
+    // gamma)(o_eval + beta * K2 * z_challenge + gamma) * alpha z(X)
     fn compute_linearizer_identity_range_check(
         &self,
-        (a_eval, b_eval, c_eval, d_eval): (
+        (a_eval, b_eval, o_eval, d_eval): (
             &BlsScalar,
             &BlsScalar,
             &BlsScalar,
@@ -196,9 +196,9 @@ impl ProverKey {
         let mut a_1 = b_eval + beta_z_k1;
         a_1 += gamma;
 
-        // c_eval + beta * K2 * z_challenge + gamma
+        // o_eval + beta * K2 * z_challenge + gamma
         let beta_z_k2 = K2 * beta_z;
-        let mut a_2 = c_eval + beta_z_k2;
+        let mut a_2 = o_eval + beta_z_k2;
         a_2 += gamma;
 
         // d_eval + beta * K3 * z_challenge + gamma
@@ -210,17 +210,17 @@ impl ProverKey {
         a *= a_2;
         a *= a_3;
         a *= alpha; // (a_eval + beta * z_challenge + gamma)(b_eval + beta * K1 *
-                    // z_challenge + gamma)(c_eval + beta * K2 * z_challenge + gamma)(d_eval
+                    // z_challenge + gamma)(o_eval + beta * K2 * z_challenge + gamma)(d_eval
                     // + beta * K3 * z_challenge + gamma) * alpha
         z_poly * &a // (a_eval + beta * z_challenge + gamma)(b_eval + beta * K1
-                    // * z_challenge + gamma)(c_eval + beta * K2 * z_challenge +
+                    // * z_challenge + gamma)(o_eval + beta * K2 * z_challenge +
                     // gamma) * alpha z(X)
     }
     // -(a_eval + beta * sigma_1 + gamma)(b_eval + beta * sigma_2 + gamma)
-    // (c_eval + beta * sigma_3 + gamma) * beta *z_eval * alpha^2 * Sigma_4(X)
+    // (o_eval + beta * sigma_3 + gamma) * beta *z_eval * alpha^2 * Sigma_4(X)
     fn compute_linearizer_copy_range_check(
         &self,
-        (a_eval, b_eval, c_eval): (&BlsScalar, &BlsScalar, &BlsScalar),
+        (a_eval, b_eval, o_eval): (&BlsScalar, &BlsScalar, &BlsScalar),
         z_eval: &BlsScalar,
         sigma_1_eval: &BlsScalar,
         sigma_2_eval: &BlsScalar,
@@ -238,9 +238,9 @@ impl ProverKey {
         let mut a_1 = b_eval + beta_sigma_2;
         a_1 += gamma;
 
-        // c_eval + beta * sigma_3 + gamma
+        // o_eval + beta * sigma_3 + gamma
         let beta_sigma_3 = beta * sigma_3_eval;
-        let mut a_2 = c_eval + beta_sigma_3;
+        let mut a_2 = o_eval + beta_sigma_3;
         a_2 += gamma;
 
         let beta_z_eval = beta * z_eval;
@@ -248,10 +248,10 @@ impl ProverKey {
         let mut a = a_0 * a_1 * a_2;
         a *= beta_z_eval;
         a *= alpha; // (a_eval + beta * sigma_1 + gamma)(b_eval + beta * sigma_2 +
-                    // gamma)(c_eval + beta * sigma_3 + gamma) * beta * z_eval * alpha
+                    // gamma)(o_eval + beta * sigma_3 + gamma) * beta * z_eval * alpha
 
         s_sigma_4_poly * &-a // -(a_eval + beta * sigma_1 + gamma)(b_eval +
-                             // beta * sigma_2 + gamma) (c_eval + beta *
+                             // beta * sigma_2 + gamma) (o_eval + beta *
                              // sigma_3 + gamma) * beta * z_eval * alpha^2 *
                              // Sigma_4(X)
     }