+++ src/proof_system/widget/range/proverkey.rs
@@ -34,7 +34,7 @@ impl ProverKey {
         range_separation_challenge: &BlsScalar,
         a_w_i: &BlsScalar,
         b_w_i: &BlsScalar,
-        c_w_i: &BlsScalar,
+        o_w_i: &BlsScalar,
         d_w_i: &BlsScalar,
         d_w_i_next: &BlsScalar,
     ) -> BlsScalar {
@@ -45,11 +45,11 @@ impl ProverKey {
         let kappa_sq = kappa.square();
         let kappa_cu = kappa_sq * kappa;
 
-        // Delta([c(X) - 4 * d(X)]) + Delta([b(X) - 4 * c(X)]) + Delta([a(X) - 4
+        // Delta([o(X) - 4 * d(X)]) + Delta([b(X) - 4 * o(X)]) + Delta([a(X) - 4
         // * b(X)]) + Delta([d(Xg) - 4 * a(X)]) * Q_Range(X)
         //
-        let b_1 = delta(c_w_i - four * d_w_i);
-        let b_2 = delta(b_w_i - four * c_w_i) * kappa;
+        let b_1 = delta(o_w_i - four * d_w_i);
+        let b_2 = delta(b_w_i - four * o_w_i) * kappa;
         let b_3 = delta(a_w_i - four * b_w_i) * kappa_sq;
         let b_4 = delta(d_w_i_next - four * a_w_i) * kappa_cu;
         (b_1 + b_2 + b_3 + b_4) * q_range_i * range_separation_challenge
@@ -60,7 +60,7 @@ impl ProverKey {
         range_separation_challenge: &BlsScalar,
         a_eval: &BlsScalar,
         b_eval: &BlsScalar,
-        c_eval: &BlsScalar,
+        o_eval: &BlsScalar,
         d_eval: &BlsScalar,
         d_next_eval: &BlsScalar,
     ) -> Polynomial {
@@ -71,11 +71,11 @@ impl ProverKey {
         let kappa_sq = kappa.square();
         let kappa_cu = kappa_sq * kappa;
 
-        // Delta([c_eval - 4 * d_eval]) + Delta([b_eval - 4 * c_eval]) +
+        // Delta([o_eval - 4 * d_eval]) + Delta([b_eval - 4 * o_eval]) +
         // Delta([a_eval - 4 * b_eval]) + Delta([d_next_eval - 4 * a_eval]) *
         // Q_Range(X)
-        let b_1 = delta(c_eval - four * d_eval);
-        let b_2 = delta(b_eval - four * c_eval) * kappa;
+        let b_1 = delta(o_eval - four * d_eval);
+        let b_2 = delta(b_eval - four * o_eval) * kappa;
         let b_3 = delta(a_eval - four * b_eval) * kappa_sq;
         let b_4 = delta(d_next_eval - four * a_eval) * kappa_cu;